requirements/base.txt
Django==1.8.6
djangorestframework==3.3.2
